Move some dataReducer methods to unNormalize
