+++ src/util/dataReducers.js
@@ -1,59 +1,8 @@
 /**
  * Created by Tõnis Kasekamp on 18.12.2017.
  */
-import {CLASSIFICATION, REGRESSION, SPLIT_SINGLE} from '../reference';
+import {CLASSIFICATION, REGRESSION} from '../reference';
 
-const logOrDefault = (logsById, log) => (logsById[log] ? logsById[log].name : '');
-export const mergeSplitWithLogName = (splitsById, logsById) => {
-  return Object.values(splitsById).map((split) => ({
-    ...split,
-    originalLogName: logOrDefault(logsById, split.original_log),
-    testLogName: logOrDefault(logsById, split.test_log),
-    trainingLogName: logOrDefault(logsById, split.training_log)
-  }));
-};
-
-export const splitsToLabel = (splitsById, logsById) => {
-  return Object.values(splitsById).map((split) => {
-    return {value: split.id, label: splitToLabel(logsById, split)};
-  });
-};
-
-/* eslint-disable max-len */
-export const splitToLabel = (logsById, split) => {
-  let label;
-  if (split.type === SPLIT_SINGLE) {
-    label = `Split #${split.id}, ${logOrDefault(logsById, split.original_log)}`;
-  } else {
-    label = `Split #${split.id}, logs ${logOrDefault(logsById, split.training_log)} and ${logOrDefault(logsById, split.test_log)}`;
-  }
-  return label;
-};
-
-export const splitsToString = (splits) => {
-  return splits.map((split) => {
-    return {value: split.id, label: splitToString(split)};
-  });
-};
-
-export const splitToString = (split) => {
-  let label;
-  if (split.type === SPLIT_SINGLE) {
-    label = `Split #${split.id}, ${split.original_log.name}`;
-  } else {
-    label = `Split #${split.id}, logs ${split.training_log.name} and ${split.test_log.name}`;
-  }
-  return label;
-};
-
-export const mapJobs = (logsById, splitsById, jobsById) => {
-  return Object.values(jobsById).map((job) => {
-    return {
-      ...job,
-      splitName: splitToLabel(logsById, splitsById[job.split_id])
-    };
-  });
-};
 
 export const jobToValidationTable = (job) => {
   const kmeans = job.config.kmeans;
@@ -64,7 +13,7 @@ export const jobToValidationTable = (job) => {
       encoding: job.config.encoding,
       clustering: job.config.clustering,
       method: job.config.method,
-      splitName: splitToString(job.split),
+      splitName: '',
       prefix_length: job.config.prefix_length,
       padding: job.config.padding,
       hyperopt: job.config.hyperopt,
@@ -80,7 +29,7 @@ export const jobToValidationTable = (job) => {
       encoding: job.config.encoding,
       clustering: job.config.clustering,
       method: job.config.method,
-      splitName: splitToString(job.split),
+      splitName: '',
       prefix_length: job.config.prefix_length,
       padding: job.config.padding,
       hyperopt: job.config.hyperopt,
@@ -93,7 +42,7 @@ export const jobToValidationTable = (job) => {
     return {
       id: job.id,
       encoding: job.config.encoding,
-      splitName: splitToString(job.split),
+      splitName: '',
       prefix_length: job.config.prefix_length,
       padding: job.config.padding,
       label: job.config.label,
